requirements.txt
tensorflow==2.15.0
keras
pillow
numpy


fastapi
python-multipart
uvicorn[standard]